pricing constant updated
